fix: output when blockchain.info fails to respond.
The module we were relying on before for getting the JSON from blockchain.info (`get-json`) doesn't properly
call the callback we pass in if the response status code isn't 200, which means that we have no way of knowing
if the response is ever received, and thus, no way to output an error. Because of that, this commit changes to
using the `request` module directly.